+++ packages/extension/src/companion/CompanionContextMenu.tsx
@@ -4,9 +4,10 @@ import FlagIcon from '@dailydotdev/shared/src/components/icons/Flag';
 import FeedbackIcon from '@dailydotdev/shared/src/components/icons/Feedback';
 import EyeIcon from '@dailydotdev/shared/src/components/icons/Eye';
 import { Item, Menu } from '@dailydotdev/react-contexify';
-import ReportPostModal, {
+import {
+  ReportPostModal,
   ReportedCallback,
-} from '@dailydotdev/shared/src/components/modals/ReportPostModal';
+} from '@dailydotdev/shared/src/components/modals/report';
 import { PostBootData } from '@dailydotdev/shared/src/lib/boot';
 import { useToastNotification } from '@dailydotdev/shared/src/hooks/useToastNotification';
 import { ShareBookmarkProps } from '@dailydotdev/shared/src/components/post/PostActions';
