+++ packages/shared/src/components/modals/report/index.ts
@@ -1,0 +1,2 @@
+export * from './ReportCommentModal';
+export * from './ReportPostModal';
